requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography>=41.0.0